Migrates pull.fitbit to new plugin layout
